Add WithVersionIntent option setter & move VersionIntent to temporal pkg
